Update hmf.py
The check valid url method is doign a GET request which is causing large files to be downloaded in the background. Kodiplayer is also downloading the same file to diaplay, so end up with two processes pulling the same large file causing wasted bandwidth.
Doing a HEAD request achieves the valid file check while not downloading the entire file